Fix HashRouter
